refactor: removed transport trait for now
