--- jarust_transport_next/src/interfaces/mod.rs
@@ -1,2 +1,0 @@
-mod websocket;
-pub mod websocket_transport;
