+++ jarust_transport_next/src/lib.rs
@@ -1,12 +1,10 @@
 pub mod error;
 pub mod handle_msg;
-pub mod interfaces;
 pub mod japrotocol;
 pub mod jatransport;
 pub mod prelude;
 pub mod respones;
 pub mod transaction_gen;
-pub mod transport;
 
 mod demuxer;
 mod napmap;
